Refactored HoodieFileScanRDD to make it compatible with all Spark up to 3.3
