+++ hudi-spark-datasource/hudi-spark3.1.x/src/main/scala/org/apache/hudi/Spark31HoodieFileScanRDD.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.execution.datasources
+package org.apache.hudi
 
 import org.apache.hudi.HoodieUnsafeRDD
 import org.apache.spark.sql.SparkSession
@@ -24,11 +24,10 @@ import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.AttributeReference
 import org.apache.spark.sql.types.StructType
 
-class Spark33HoodieFileScanRDD(@transient private val sparkSession: SparkSession,
-                        readFunction: PartitionedFile => Iterator[InternalRow],
-                        @transient filePartitions: Seq[FilePartition],
-                        readDataSchema: StructType, metadataColumns: Seq[AttributeReference] = Seq.empty)
-  extends FileScanRDD(sparkSession, readFunction, filePartitions, readDataSchema, metadataColumns)
+class Spark31HoodieFileScanRDD(@transient private val sparkSession: SparkSession,
+                               readFunction: PartitionedFile => Iterator[InternalRow],
+                               @transient filePartitions: Seq[FilePartition])
+  extends FileScanRDD(sparkSession, readFunction, filePartitions)
     with HoodieUnsafeRDD {
 
   override final def collect(): Array[InternalRow] = super[HoodieUnsafeRDD].collect()
